% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isurvLRT.R, R/survLRT.R
\name{checkLinearityPairwise}
\alias{checkLinearityPairwise}
\title{Test for pairwaise epistasis type}
\usage{
checkLinearityPairwise(gene1, gene2, alive, GlastFollowUp)

checkLinearityPairwise(gene1, gene2, alive, GlastFollowUp)
}
\arguments{
\item{gene1}{character; gene A alterations}

\item{gene2}{character; gene B alterations}

\item{alive}{status to indicate alive or dead 1 - alive, 0 - dead}

\item{GlastFollowUp}{the probablity of surviving more than the corresponding number of months}
}
\description{
Test for pairwaise epistasis type

Test for pairwaise epistasis type
}
