% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprSL.R
\name{flatCorMatrix}
\alias{flatCorMatrix}
\title{Formatting the correlation matrix into a table with 4 columns containing :
Column 1 : row names (variable 1 for the correlation test)
Column 2 : clumn names (variable 2 for the correlation test)
Column 3 : the correlation coefficients
Column 4 : the p-values of the correlation}
\usage{
flatCorMatrix(cormatrix, pmatrix)
}
\arguments{
\item{cormatrix}{matrix of the correlation coefficients}

\item{pmatrix}{matrix of the correlation p-values}
}
\description{
Formatting the correlation matrix into a table with 4 columns containing :
Column 1 : row names (variable 1 for the correlation test)
Column 2 : clumn names (variable 2 for the correlation test)
Column 3 : the correlation coefficients
Column 4 : the p-values of the correlation
}
