% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gene_pairs.R
\name{genePairCombination}
\alias{genePairCombination}
\title{Prepare pair of genes to be tested}
\usage{
genePairCombination(geneA, geneB)
}
\arguments{
\item{geneA}{character; list of gene A}

\item{geneB}{character; list of gene B}
}
\value{
data frame
}
\description{
Prepare pair of genes to be tested
}
