% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isurvLRT.R
\name{icalculatePairGenotype}
\alias{icalculatePairGenotype}
\title{Calculate genotypes for pair of genes}
\usage{
icalculatePairGenotype(gene1, gene2, data, expression, thr)
}
\arguments{
\item{gene1}{character; gene A}

\item{gene2}{character; gene B}

\item{data}{data.frame; gene alteration with survival status}

\item{expression}{data.frame; gene expression in tumor samples. The columns correspond to genes and the rows to tumor samples.}

\item{thr}{expression threshold}
}
\description{
Calculate genotypes for pair of genes
}
