% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survLRT.R
\name{inputDataSurvLRT}
\alias{inputDataSurvLRT}
\title{Prepare alteration data with survival for survLRT and isurvLRT test}
\usage{
inputDataSurvLRT(alt, survival)
}
\arguments{
\item{alt}{data.frame gene alteration}

\item{survival}{data.frame input data with survival information}
}
\value{
data frame
}
\description{
Prepare alteration data with survival for survLRT and isurvLRT test
}
