% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isurvLRT.R
\name{isurvLRT}
\alias{isurvLRT}
\title{Test the pairs of genes}
\usage{
isurvLRT(gene.pair, alt, survival, expression, cores = 1)
}
\arguments{
\item{gene.pair}{data.frame; pair of genes to be tested}

\item{alt}{data.frame; gene alterations in tumor samples. The columns correspond to genes and the rows to tumor samples.}

\item{survival}{data.frame; survival status 1 - alive, 0 - dead}

\item{expression}{data.frame; gene expression in tumor samples. The columns correspond to genes and the rows to tumor samples.}

\item{cores}{number of cores to be used; default 1}
}
\value{
data frame
}
\description{
Test the pairs of genes
}
