% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isurvLRT.R
\name{isurvLRTPair}
\alias{isurvLRTPair}
\title{Test the pair of genes}
\usage{
isurvLRTPair(gene.pair, data, expression)
}
\arguments{
\item{gene.pair}{data.frame; pairs of genes to be tested. First column - gene A, second column - gene B}

\item{data}{data data.frame; gene alteration with survival status}

\item{expression}{data.frame; gene expression in tumor samples. The columns correspond to genes and the rows to tumor samples.}
}
\value{
data frame
}
\description{
Test the pair of genes
}
