% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survLRT.R
\name{LRT}
\alias{LRT}
\title{Calculate likelihood ratio test}
\usage{
LRT(S, H)
}
\arguments{
\item{S}{count of patients in group with a specific mutational and survival status}

\item{H}{precalculated sums of logs of survival function values for patients within groups with specific mutations}
}
\description{
Calculate likelihood ratio test
}
