% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoF.R
\name{runWilcoxonFor2Genes}
\alias{runWilcoxonFor2Genes}
\title{Calculate Wicoxon test for 2 genes}
\usage{
runWilcoxonFor2Genes(gene.pair, alt, expression)
}
\arguments{
\item{gene.pair}{data.frame; pairs of genes to be tested. First column - gene A, second column - gene B}

\item{alt}{data.frame; gene alterations in tumor samples. The columns correspond to genes and the rows to tumor samples.}

\item{expression}{data.frame; gene expression in tumor samples. The columns correspond to genes and the rows to tumor samples.}
}
\description{
Calculate Wicoxon test for 2 genes
}
