% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoF.R
\name{SoF}
\alias{SoF}
\title{Survival of the fittest (SoF)}
\usage{
SoF(gene.pair, alt, expression, cores = 1)
}
\arguments{
\item{gene.pair}{data.frame; pairs of genes to be tested. First column - gene A, second column - gene B}

\item{alt}{data.frame; gene alterations in tumor samples. The columns correspond to genes and the rows to tumor samples.}

\item{expression}{data.frame; gene expression in tumor samples. The columns correspond to genes and the rows to tumor samples.}

\item{cores}{number of cores to be used; default 1}
}
\description{
Survival of the fittest (SoF)
}
