% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survLRT.R
\name{survLRT}
\alias{survLRT}
\title{Test the pair of genes}
\usage{
survLRT(gene.pair, alt, survival, cores = 1)
}
\arguments{
\item{gene.pair}{pairs of genes to be tested}

\item{alt}{data.frame; gene alterations in tumor samples. The columns correspond to genes and the rows to tumor samples.}

\item{survival}{data.frame; survival status 1 - alive, 0 - dead}

\item{cores}{number of cores to be used; default 1}
}
\value{
data frame
}
\description{
Test the pair of genes
}
