% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isurvLRT.R
\name{threshold}
\alias{threshold}
\title{Calcultate threshold}
\usage{
threshold(gene.pair, expression)
}
\arguments{
\item{gene.pair}{pair of genes to be tested}

\item{expression}{data.frame; gene expression in tumor samples. The columns correspond to genes and the rows to tumor samples.}
}
\value{
data frame
}
\description{
Calcultate threshold
}
