% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprSL.R
\name{corrExprGenePair}
\alias{corrExprGenePair}
\title{Get the correlation matrix of gene expression for pair of genes}
\usage{
corrExprGenePair(gene.pair, expression)
}
\arguments{
\item{gene.pair}{data.frame; pairs of genes to be tested. First column - gene A, second column - gene B}

\item{expression}{data.frame; gene expression in tumor samples. The columns correspond to genes and the rows to tumor samples.}
}
\description{
Get the correlation matrix of gene expression for pair of genes
}
