% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isurvLRT.R
\name{irunFor2Genes}
\alias{irunFor2Genes}
\title{Run isurvLRT for gene pairs}
\usage{
irunFor2Genes(gene.pair, data, expression, thr)
}
\arguments{
\item{gene.pair}{pair of genes to be tested}

\item{data}{data.frame; gene alteration with survival status}

\item{expression}{data.frame; gene expression in tumor samples. The columns correspond to genes and the rows to tumor samples.}

\item{thr}{expression threshold}
}
\value{
result table
}
\description{
Run isurvLRT for gene pairs
}
