% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survLRT.R
\name{runFor2Genes}
\alias{runFor2Genes}
\title{Run SurvLRT for gene pairs}
\usage{
runFor2Genes(gene.pair, data)
}
\arguments{
\item{gene.pair}{data.frame; pairs of genes to be tested. First column - gene A, second column - gene B}

\item{data}{data data.frame; gene alteration with survival status}
}
\value{
result table
}
\description{
Run SurvLRT for gene pairs
}
